separate live exchange tests
